Number Guessing Game
We use the std::rand() function to generate a random integer.
To ensure different random numbers each time we run the program, we seed the random number generator using std::srand(std::time(nullptr)).
The expression std::rand() % 100 + 1 generates a random number between 1 and 100.
We prompt the user to guess the secret number.
The user’s guess is stored in the userGuess variable.
We keep track of the number of attempts using the attempts variable.
If the user’s guess is lower than the secret number, we print “Too low! Try a higher number.”
If the guess is higher, we print “Too high! Try a lower number.”
When the user guesses correctly, we congratulate them and display the secret number along with the number of attempts.
The do-while loop continues until the user guesses the correct number (i.e., userGuess == secretNumber).